Extract shared pool bounds aggregation helper
Move `aggregate_parallel` (and its `extract_metric_bounds` helper) out of
the battery bounds tracker into a `bounds_aggregation` module so the
upcoming PV bounds tracker can reuse the same parallel-combination logic.

No behavior change.
