+++ src/microgrid/battery_bounds_tracker.rs
@@ -19,15 +19,13 @@
 //!   aggregated bounds are intersected.
 //! * Groups within a pool are in parallel — their bounds are added together.
 
-use std::collections::HashMap;
 use std::marker::PhantomData;
 
 use tokio::sync::broadcast;
 
 use crate::bounds::{combine_parallel_sets, intersect_bounds_sets};
-use crate::client::proto::common::{
-    metrics::Bounds as PbBounds, microgrid::electrical_components::ElectricalComponentTelemetry,
-};
+use crate::client::proto::common::metrics::Bounds as PbBounds;
+use crate::microgrid::bounds_aggregation::aggregate_parallel;
 use crate::microgrid::telemetry_tracker::battery_pool_telemetry_tracker::BatteryPoolSnapshot;
 use crate::{Bounds, metric::Metric};
 
@@ -106,39 +104,6 @@ where
                 combine_parallel_sets(&acc, &group_bounds)
             })
     }
-}
-
-/// Combines the bounds of every component in the map as if they were wired
-/// in parallel. Components that don't report the metric `M` are skipped.
-fn aggregate_parallel<M: Metric>(
-    components: &HashMap<u64, ElectricalComponentTelemetry>,
-) -> Vec<Bounds<M::QuantityType>>
-where
-    Bounds<M::QuantityType>: From<PbBounds>,
-{
-    components
-        .values()
-        .filter_map(extract_metric_bounds::<M>)
-        .fold(Vec::new(), |acc, bounds| {
-            combine_parallel_sets(&acc, &bounds)
-        })
-}
-
-fn extract_metric_bounds<M: Metric>(
-    telemetry: &ElectricalComponentTelemetry,
-) -> Option<Vec<Bounds<M::QuantityType>>>
-where
-    Bounds<M::QuantityType>: From<PbBounds>,
-{
-    telemetry.metric_samples.iter().find_map(|sample| {
-        (sample.metric == M::METRIC as i32).then(|| {
-            sample
-                .bounds
-                .iter()
-                .map(|b| Bounds::from(*b))
-                .collect::<Vec<_>>()
-        })
-    })
 }
 
 #[cfg(test)]
